// src/domain/enums/EscrowStatus.ts
export enum EscrowStatus {
  NOT_STARTED = 'NOT_STARTED',
  AWAITING_EXTERNAL_PAYMENT = 'AWAITING_EXTERNAL_PAYMENT',
  EXTERNAL_PAYMENT_AUTHORIZED = 'EXTERNAL_PAYMENT_AUTHORIZED',
  EXTERNAL_PAYMENT_CAPTURED = 'EXTERNAL_PAYMENT_CAPTURED',
  FUNDS_RELEASED_TO_OWNER = 'FUNDS_RELEASED_TO_OWNER',
  DISPUTED = 'DISPUTED',
  REFUNDED = 'REFUNDED',
}